requirements.txt
requests==2.28.1
setuptools==63.2.0
wheel>=0.37.1
PyYAML>=6.0
jsonschema>=4.6.2
marshmallow>=3.17.0
pytest>=7.1.2
pytest-html-reporter>=0.2.9
pytest-ordering>=0.6
sphinx==4.3.2
myst-parser==0.18.0
sphinx-rtd-theme==1.0.0
sphinxcontrib-napoleon==0.7
rst2pdf==0.99
streamlit==1.10.0
